+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## Unreleased
+## [1.9.0] - 2020-08-17
 
 - Omitted binary body in FullHttpMessageFormatter. `[binary stream omitted]` will be shown instead.